refactor: zlib module structure
