+++ src/modules/commands/zlib/compress/index.js
@@ -3,13 +3,13 @@ import { pipeline } from 'stream/promises';
 import { createWriteStream, createReadStream } from 'fs';
 import { createBrotliCompress } from 'zlib';
 
-import { appData } from '../../appData.js';
-import { errorHandler } from '../../errorHandler.js';
+import { appData } from '../../../appData.js';
+import { errorHandler } from '../../../errorHandler.js';
 
-import { printMessage } from '../../../utils/printMessage.js';
-import { exists } from '../../../utils/exists.js';
+import { printMessage } from '../../../../utils/printMessage.js';
+import { exists } from '../../../../utils/exists.js';
 
-import { BROTLI_EXTNAME } from '../../../constants/index.js';
+import { BROTLI_EXTNAME } from '../../../../constants/index.js';
 
 export const compress = async (pathToFile, pathToDir) => {
     const originalPath = resolve(appData.cwd, pathToFile);
