+++ src/modules/commands/zlib/decompress/helpers/resolveExtname.js
@@ -1,3 +1,3 @@
-import { BROTLI_EXTNAME } from '../../../../constants/index.js';
+import { BROTLI_EXTNAME } from '../../../../../constants/index.js';
 
 export const resolveExtname = pathToFile => pathToFile.endsWith(BROTLI_EXTNAME) ? pathToFile : pathToFile + BROTLI_EXTNAME;
