+++ src/modules/commands/zlib/decompress/index.js
@@ -5,13 +5,13 @@ import { createBrotliDecompress } from 'zlib';
 
 import { resolveExtname } from './helpers/resolveExtname.js';
 
-import { appData } from '../../appData.js';
-import { errorHandler } from '../../errorHandler.js';
+import { appData } from '../../../appData.js';
+import { errorHandler } from '../../../errorHandler.js';
 
-import { printMessage } from '../../../utils/printMessage.js';
-import { exists } from '../../../utils/exists.js';
+import { printMessage } from '../../../../utils/printMessage.js';
+import { exists } from '../../../../utils/exists.js';
 
-import { BROTLI_EXTNAME } from '../../../constants/index.js';
+import { BROTLI_EXTNAME } from '../../../../constants/index.js';
 
 export const decompress = async (pathToFile, pathToDir) => {
     const originalPath = resolve(appData.cwd, resolveExtname(pathToFile));
