+++ src/modules/commands/zlib/index.js
@@ -1,0 +1,2 @@
+export { compress } from './compress/index.js';
+export { decompress } from './decompress/index.js';
